requirements.txt
gradio>=4.0
python-dotenv>=0.21.0
pydantic-settings>=1.0.0
pytest>=7.0
pandas>=1.5
numpy>=1.24
sentence-transformers>=2.2
chromadb>=0.3
pdfminer.six>=20221105
openai>=0.27
